requirements.txt
boto3==1.15.2
botocore==1.18.2
django-cors-headers==3.5.0
django-ses==1.0.3
future==0.18.2
jmespath==0.10.0
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
asgiref==3.2.10
Django==3.1.1
django-phone-field==1.8.1
djangorestframework==3.11.1
pytz==2020.1
sqlparse==0.3.1
